+++ src/main/java/luminous/StudentForum/repository/CategoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import luminous.StudentForum.model.Category;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Short> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
     Category findByCname(String cname);
 }
